requirements.txt
pyserial
pyubx2
pyYAML
